Refactor module object in powerhub.ps1; auto-update
We now use a more idiomatic PSObject for a Module on the powershell
side. The modules list is now loaded in an extra web request so it can
be automatically updated. When the module list is updated, the module
directory will be scanned again on the server side to have the most
recent state.