Remove eventlet dependency
This dependency was only needed because socketio was importing it at the
top-level. Starting with version 5.8, socketio is lazy-loading it, and
since we don't need the feature it is being loaded for (message queues),
we don't need eventlet anymore. That's good because eventlet has an
annoying bug which makes it incompatible with dnspython.

Close #63.